Rename workshop.md to index.md
for the time being, the index.md file displays the contents for the website. therefore I renamed the workshop.md file to index.md. If we later want to create a different website with more than one page, the contents of index.md should be saved into a new file; possibly workshop.md or something similar.